fixes for text on raspberrypi

+++ src/sg/base.cpp
@@ -88,7 +88,7 @@ static void add_text( vertex_buffer_t * buffer, texture_font_t * font,
             float t0 = glyph->t0;
             float s1 = glyph->s1;
             float t1 = glyph->t1;
-            GLuint indices[6] = {0,1,2, 0,2,3};
+            GLushort indices[6] = {0,1,2, 0,2,3};
             vertex_t vertices[4] = { { x0,y0,0,  s0,t0,  r,g,b,a },
                                      { x0,y1,0,  s0,t1,  r,g,b,a },
                                      { x1,y1,0,  s1,t1,  r,g,b,a },
