+++ src/sg/fonts/vertex-buffer.h
@@ -231,7 +231,7 @@ typedef struct
  */
   void
   vertex_buffer_push_back_indices ( vertex_buffer_t *self,
-                                    const GLuint * indices,
+                                    const GLushort * indices,
                                     const size_t icount );
 
 
@@ -265,7 +265,7 @@ typedef struct
   void
   vertex_buffer_insert_indices ( vertex_buffer_t *self,
                                  const size_t index,
-                                 const GLuint *indices,
+                                 const GLushort *indices,
                                  const size_t icount );
 
 
@@ -326,7 +326,7 @@ typedef struct
   size_t
   vertex_buffer_push_back( vertex_buffer_t * self,
                            const void * vertices, const size_t vcount,  
-                           const GLuint * indices, const size_t icount );
+                           const GLushort * indices, const size_t icount );
 
 
 /**
@@ -343,7 +343,7 @@ typedef struct
   vertex_buffer_insert( vertex_buffer_t * self,
                         size_t index,
                         const void * vertices, const size_t vcount,  
-                        const GLuint * indices, const size_t icount );
+                        const GLushort * indices, const size_t icount );
 
 /**
  * Erase an item from the vertex buffer.
